moved API endpoints, added auction_id url params
