+++ backend/auction/views.py
@@ -7,6 +7,8 @@ from rest_framework.views import APIView
 
 from .models import Auction
 from .serializers import AuctionSerializer
+from vehicle.models import Vehicle, SavedUnits, ContentType
+from user.models import Bidder
 
 
 class AuctionListApiView(APIView):
@@ -86,3 +88,82 @@ class AuctionDetailApiView(APIView):
         auction = get_object_or_404(Auction, id=auction_id)
         auction.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
+
+
+class SaveUnitApiView(APIView):
+    """
+    An endpoint to handle saving a vehicle to a bidder's saved vehicle list,
+    as well as retrieving a list of all saved vehicles
+    """
+
+    def post(self, request, **kwargs):
+        bidder_id = kwargs.get("bidder_id")
+        vehicle_id = kwargs.get("vehicle_id")
+        auction_id = kwargs.get("auction_id")
+
+        # Replace this later to fetch authentication details
+        # from headers instead of body
+        vehicle = get_object_or_404(Vehicle, id=vehicle_id)
+        auction_for_vehicle = Auction.objects.get(id=auction_id)
+        bidder = get_object_or_404(Bidder, id=bidder_id)
+        if SavedUnits.objects.filter(
+            auction_id=auction_for_vehicle, bidder_id=bidder, object_id=vehicle.id
+        ):
+            return Response(
+                {"message": "Vehicle already saved"},
+                status=status.HTTP_204_NO_CONTENT,
+            )
+
+        saved_unit = SavedUnits(
+            auction_id=auction_for_vehicle,
+            bidder_id=bidder,
+            object_id=vehicle.id,
+            content_object=vehicle,
+        )
+        saved_unit.save()
+        return Response(
+            {"message": "Vehicle saved successfully"},
+            status=status.HTTP_200_OK,
+        )
+
+    def delete(self, request, **kwargs):
+        vehicle_id = kwargs.get("vehicle_id")
+        bidder_id = kwargs.get("bidder_id")
+        auction = kwargs.get("auction_id")
+        bidder = get_object_or_404(Bidder, id=bidder_id)
+
+        saved_unit = get_object_or_404(
+            SavedUnits, object_id=vehicle_id, bidder_id=bidder, auction_id=auction
+        )
+
+        saved_unit.delete()
+
+        return Response(
+            {"message": "Saved unit deleted successfully"},
+            status=status.HTTP_200_OK,
+        )
+
+
+class GetSavedUnitApiView(APIView):
+    """
+    An endpoint to retrieve all of bidder's saved units associated with
+    a provided auction
+    """
+
+    def get(self, request, **kwargs):
+        # Replace this line to obtain user id once authentication has been implemented
+        bidder_id = kwargs.get("bidder_id")
+        auction_id = kwargs.get("auction_id")
+        bidder = get_object_or_404(Bidder, id=bidder_id)
+        auction = get_object_or_404(Auction, id=auction_id)
+
+        saved_units = SavedUnits.objects.filter(bidder_id=bidder, auction_id=auction)
+        vehicle_list = [
+            saved_unit.content_object
+            for saved_unit in saved_units
+            if isinstance(saved_unit.content_object, Vehicle)
+        ]
+
+        vehicle_data = [{"id": vehicle.id} for vehicle in vehicle_list]
+
+        return Response({"vehicles": vehicle_data}, status=status.HTTP_200_OK)
