+++ backend/user/urls.py
@@ -8,7 +8,6 @@ from .views import (
     BidderDetailApiView,
     BidderListApiView,
     BidderVerifyApiView,
-    SaveUnitApiView,
     ListBlacklisted,
     ListUnverified,
     ListVerified,
@@ -31,11 +30,6 @@ urlpatterns = [
     ),
     path("admins/", AdminListApiView.as_view(), name="admin-list"),
     path("admins/<uuid:admin_id>/", AdminDetailApiView.as_view(), name="admin-detail"),
-    path(
-        "<uuid:bidder_id>/vehicles/<uuid:vehicle_id>",
-        SaveUnitApiView.as_view(),
-        name="save-unit-to-list",
-    ),
     path("bidders/unverified", ListUnverified.as_view(), name="unverified-list"),
     path("bidders/blacklisted", ListBlacklisted.as_view(), name="blacklisted-list"),
     path("bidders/verified", ListVerified.as_view(), name="verified-list"),
